Refactor viewport helpers
