+++ webdriver/tests/bidi/browsing_context/capture_screenshot/frame.py
@@ -7,12 +7,12 @@ from tests.support.screenshot import (DEFAULT_CONTENT,
                                       OUTER_IFRAME_STYLE,
                                       INNER_IFRAME_STYLE)
 
-from . import viewport_dimensions
+from . import physical_viewport_dimensions
 
 
 @pytest.mark.asyncio
 async def test_iframe(bidi_session, top_context, inline, iframe):
-    viewport_size = await viewport_dimensions(bidi_session, top_context)
+    viewport_size = await physical_viewport_dimensions(bidi_session, top_context)
 
     iframe_content = f"{INNER_IFRAME_STYLE}{DEFAULT_CONTENT}"
     url = inline(f"{OUTER_IFRAME_STYLE}{iframe(iframe_content)}")
@@ -34,7 +34,7 @@ async def test_iframe(bidi_session, top_context, inline, iframe):
 @pytest.mark.parametrize("domain", ["", "alt"], ids=["same_origin", "cross_origin"])
 @pytest.mark.asyncio
 async def test_context_origin(bidi_session, top_context, inline, iframe, compare_png_bidi, domain):
-    expected_size = await viewport_dimensions(bidi_session, top_context)
+    expected_size = await physical_viewport_dimensions(bidi_session, top_context)
 
     initial_url = inline(f"{REFERENCE_STYLE}{REFERENCE_CONTENT}")
     await bidi_session.browsing_context.navigate(context=top_context["context"],
