+++ webdriver/tests/bidi/input/perform_actions/pointer_origin.py
@@ -8,9 +8,7 @@ from tests.support.helpers import filter_dict
 
 from .. import get_events
 from . import (
-    get_element_rect,
     get_inview_center_bidi,
-    get_viewport_rect,
     remote_mapping_to_dict,
 )
